1. SwiftUICore module is no longer available, so removed it 2. UI and api integration modifications in connect email sync 3. Review bottom sheet UI alignment issue fixed

+++ Subzillo/Features/AddSubscription/ConnectEmail/ConnectedEmailsListView/SwipeableEmailRow.swift
@@ -8,8 +8,8 @@ struct SwipeableMailRow: View {
     @Binding var activeCardId       : String?
     @Binding var isScrollDisabled   : Bool
     let onDelete                    : () -> Void
-    let onSync                      : () -> Void
-    let onView                      : () -> Void
+    let onSync                      : (Int) -> Void
+    let onView                      : (Int) -> Void
     let onDownloadLogs              : () -> Void
     @State private var offset       : CGFloat = 0
     @State private var isSwiped     : Bool = false
